Add the infrastructure for boolean[]-backed images
Previously, one could extend the type system to have a NativeBoolType
(plus BoolAccess plus BoolArray) externally, and use it with ArrayImg
and/or PlanarImg. However, as of ImgLib2 v5, the NativeTypeFactory was
locked down tightly, as a final class with only a private constructor.
Hence, creating a NativeTypeFactory that produced NativeBoolType
instances became impossible.

This commit addresses the issue by introducing the needed classes to
work with boolean arrays in the same way as other primitive arrays.